removed blueprint library
